backend/pyproject.toml
[tool.poetry]
name = "bb-backend"
version = "0.1.0"
description = ""
authors = ["Fredrik Stockman <fredrik@versionlens.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.4"
django-ninja = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
